Sensors/UM100SPI: Set default antenna offset, and fixed bug where value should have been a double not integer.

+++ src/Sensors/UM100SPI/UM100Hal.hpp
@@ -91,7 +91,7 @@ namespace Sensors
       // External data length in bits.
       uint32_t extDataLenInBits;
       // Antenna offset
-      uint16_t antenna_offset;
+      double antenna_offset;
     };
 
     enum Role
